Separation of file management methods
