+++ Assets/System/FileManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4741e898ae9855f46a1279d9de2aa392
+timeCreated: 1456770071
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
